# PE: Get actual checksum speedup ##bin